+++ js/utils.js
@@ -1,0 +1,19 @@
+// random integer
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+// random element
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+// id generator
+const createIdGenerator = () => {
+  let lastGeneratedId = 0;
+  return () => {
+    lastGeneratedId += 1;
+    return lastGeneratedId;
+  };
+};
+
+export{getRandomInteger, getRandomArrayElement, createIdGenerator};
